+++ CHANGELOG.md
@@ -13,6 +13,15 @@ and this project adheres to [Semantic Versioning][].
 ### Removed
 -->
 
+## [0.2.1][] - 2026-01-21
+
+### Changed
+
+* `make` command now uses original text (msgid) as fallback when translation
+  is empty, has `notranslate` flag, or entry is missing in PO file.
+
+[0.2.1]: https://github.com/WoozyMasta/dayz-stringtable/compare/v0.2.0...v0.2.1
+
 ## [0.2.0][] - 2026-01-21
 
 ### Added